fix 修复 部门实体与 bo vo 排序字段类型不匹配

+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/domain/bo/SysDeptBo.java
@@ -42,7 +42,7 @@ public class SysDeptBo extends BaseEntity {
      * 显示顺序
      */
     @NotNull(message = "显示顺序不能为空")
-    private Long orderNum;
+    private Integer orderNum;
 
     /**
      * 负责人
